RL/requirements.txt
asttokens                2.4.1
backcall                 0.2.0
comm                     0.2.2
contourpy                1.1.1
cycler                   0.12.1
debugpy                  1.6.7
decorator                5.1.1
dirichletprocess         0.2.0
entrypoints              0.4
executing                2.1.0
filelock                 3.14.0
fonttools                4.55.0
fsspec                   2024.3.1
importlib_resources      6.4.5
ipykernel                6.29.5
ipython                  8.12.0
jedi                     0.19.1
Jinja2                   3.1.4
joblib                   1.4.2
jupyter-client           7.3.4
jupyter_core             5.7.2
kafka-python             2.0.2
kiwisolver               1.4.7
MarkupSafe               2.1.5
matplotlib               3.7.5
matplotlib-inline        0.1.7
mpmath                   1.3.0
nest_asyncio             1.6.0
networkx                 3.1
numpy                    1.24.4
nvidia-cublas-cu12       12.1.3.1
nvidia-cuda-cupti-cu12   12.1.105
nvidia-cuda-nvrtc-cu12   12.1.105
nvidia-cuda-runtime-cu12 12.1.105
nvidia-cudnn-cu12        8.9.2.26
nvidia-cufft-cu12        11.0.2.54
nvidia-curand-cu12       10.3.2.106
nvidia-cusolver-cu12     11.4.5.107
nvidia-cusparse-cu12     12.1.0.106
nvidia-nccl-cu12         2.20.5
nvidia-nvjitlink-cu12    12.4.127
nvidia-nvtx-cu12         12.1.105
packaging                24.1
pandas                   2.0.3
parso                    0.8.4
pexpect                  4.9.0
pickleshare              0.7.5
pillow                   10.3.0
pip                      23.3.1
platformdirs             4.3.6
prompt_toolkit           3.0.47
psutil                   5.9.0
ptyprocess               0.7.0
pure_eval                0.2.3
Pygments                 2.18.0
pyparsing                3.1.4
python-dateutil          2.9.0
pytz                     2024.2
pyzmq                    25.1.2
scikit-learn             1.3.2
scipy                    1.10.1
setuptools               68.2.2
six                      1.16.0
stack-data               0.6.2
sympy                    1.12
threadpoolctl            3.5.0
torch                    2.3.0
torchaudio               2.3.0
torchvision              0.18.0
tornado                  6.1
tqdm                     4.67.1
traitlets                5.14.3
triton                   2.3.0
typing_extensions        4.11.0
tzdata                   2024.2
wcwidth                  0.2.13
wheel                    0.43.0
zipp                     3.20.2